+++ R/demean.R
@@ -11,8 +11,10 @@
 #' @param x A data frame.
 #' @param select Character vector (or formula) with names of variables to select
 #'   that should be group- and de-meaned.
-#' @param by Character vector (or formula) with the name of the variable that
-#'   indicates the group- or cluster-ID.
+#' @param by Character vector (or formula) with the name of the variable(s) that
+#'   indicates the group- or cluster-ID. For cross-classified designs, `by` can
+#'   also identify two or more variables as group- or cluster-IDs. See also
+#'   section _De-meaning for cross-classified designs_ below.
 #' @param center Method for centering. `demean()` always performs
 #'   mean-centering, while `degroup()` can use `center = "median"` or
 #'   `center = "mode"` for median- or mode-centering, and also `"min"`
@@ -34,136 +36,151 @@
 #' de-meaned variable) by default.
 #'
 #' @seealso If grand-mean centering (instead of centering within-clusters)
-#'   is required, see [center()]. See [`performance::check_heterogeneity_bias()`]
+#'   is required, see [`center()`]. See [`performance::check_heterogeneity_bias()`]
 #'   to check for heterogeneity bias.
 #'
-#' @details
+#' @section Heterogeneity Bias:
 #'
-#'   \subsection{Heterogeneity Bias}{
-#'     Mixed models include different levels of sources of variability, i.e.
-#'     error terms at each level. When macro-indicators (or level-2 predictors,
-#'     or higher-level units, or more general: *group-level predictors that
-#'     **vary** within and across groups*) are included as fixed effects (i.e.
-#'     treated as covariate at level-1), the variance that is left unaccounted for
-#'     this covariate will be absorbed into the error terms of level-1 and level-2
-#'     (\cite{Bafumi and Gelman 2006; Gelman and Hill 2007, Chapter 12.6.}):
-#'     \dQuote{Such covariates contain two parts: one that is specific to the
-#'     higher-level entity that does not vary between occasions, and one that
-#'     represents the difference between occasions, within higher-level entities}
-#'     (\cite{Bell et al. 2015}). Hence, the error terms will be correlated with
-#'     the covariate, which violates one of the assumptions of mixed models
-#'     (iid, independent and identically distributed error terms). This bias is
-#'     also called the *heterogeneity bias* (\cite{Bell et al. 2015}). To
-#'     resolve this problem, level-2 predictors used as (level-1) covariates should
-#'     be separated into their "within" and "between" effects by "de-meaning" and
-#'     "group-meaning": After demeaning time-varying predictors, \dQuote{at the
-#'     higher level, the mean term is no longer constrained by Level 1 effects,
-#'     so it is free to account for all the higher-level variance associated
-#'     with that variable} (\cite{Bell et al. 2015}).
-#'   }
+#' Mixed models include different levels of sources of variability, i.e.
+#' error terms at each level. When macro-indicators (or level-2 predictors,
+#' or higher-level units, or more general: *group-level predictors that
+#' **vary** within and across groups*) are included as fixed effects (i.e.
+#' treated as covariate at level-1), the variance that is left unaccounted for
+#' this covariate will be absorbed into the error terms of level-1 and level-2
+#' (_Bafumi and Gelman 2006; Gelman and Hill 2007, Chapter 12.6._):
+#' "Such covariates contain two parts: one that is specific to the higher-level
+#' entity that does not vary between occasions, and one that represents the
+#' difference between occasions, within higher-level entities" (_Bell et al. 2015_).
+#' Hence, the error terms will be correlated with the covariate, which violates
+#' one of the assumptions of mixed models (iid, independent and identically
+#' distributed error terms). This bias is also called the *heterogeneity bias*
+#' (_Bell et al. 2015_). To resolve this problem, level-2 predictors used as
+#' (level-1) covariates should be separated into their "within" and "between"
+#' effects by "de-meaning" and "group-meaning": After demeaning time-varying
+#' predictors, "at the higher level, the mean term is no longer constrained by
+#' Level 1 effects, so it is free to account for all the higher-level variance
+#' associated with that variable" (_Bell et al. 2015_).
 #'
-#'   \subsection{Panel data and correlating fixed and group effects}{
-#'     `demean()` is intended to create group- and de-meaned variables
-#'     for panel regression models (fixed effects models), or for complex
-#'     random-effect-within-between models (see \cite{Bell et al. 2015, 2018}),
-#'     where group-effects (random effects) and fixed effects correlate (see
-#'     \cite{Bafumi and Gelman 2006}). This can happen, for instance, when
-#'     analyzing panel data, which can lead to *Heterogeneity Bias*. To
-#'     control for correlating predictors and group effects, it is recommended
-#'     to include the group-meaned and de-meaned version of *time-varying covariates*
-#'     (and group-meaned version of *time-invariant covariates* that are on
-#'     a higher level, e.g. level-2 predictors) in the model. By this, one can
-#'     fit complex multilevel models for panel data, including time-varying
-#'     predictors, time-invariant predictors and random effects.
-#'    }
+#' @section Panel data and correlating fixed and group effects:
 #'
-#'   \subsection{Why mixed models are preferred over fixed effects models}{
-#'     A mixed models approach can model the causes of endogeneity explicitly
-#'     by including the (separated) within- and between-effects of time-varying
-#'     fixed effects and including time-constant fixed effects. Furthermore,
-#'     mixed models also include random effects, thus a mixed models approach
-#'     is superior to classic fixed-effects models, which lack information of
-#'     variation in the group-effects or between-subject effects. Furthermore,
-#'     fixed effects regression cannot include random slopes, which means that
-#'     fixed effects regressions are neglecting \dQuote{cross-cluster differences
-#'     in the effects of lower-level controls (which) reduces the precision of
-#'     estimated context effects, resulting in unnecessarily wide confidence
-#'     intervals and low statistical power} (\cite{Heisig et al. 2017}).
-#'   }
+#' `demean()` is intended to create group- and de-meaned variables for panel
+#' regression models (fixed effects models), or for complex
+#' random-effect-within-between models (see _Bell et al. 2015, 2018_), where
+#' group-effects (random effects) and fixed effects correlate (see
+#' _Bafumi and Gelman 2006_). This can happen, for instance, when analyzing
+#' panel data, which can lead to *Heterogeneity Bias*. To control for correlating
+#' predictors and group effects, it is recommended to include the group-meaned
+#' and de-meaned version of *time-varying covariates* (and group-meaned version
+#' of *time-invariant covariates* that are on a higher level, e.g. level-2
+#' predictors) in the model. By this, one can fit complex multilevel models for
+#' panel data, including time-varying predictors, time-invariant predictors and
+#' random effects.
 #'
-#'   \subsection{Terminology}{
-#'     The group-meaned variable is simply the mean of an independent variable
-#'     within each group (or id-level or cluster) represented by `by`.
-#'     It represents the cluster-mean of an independent variable. The regression
-#'     coefficient of a group-meaned variable is the *between-subject-effect*.
-#'     The de-meaned variable is then the centered version of the group-meaned
-#'     variable. De-meaning is sometimes also called person-mean centering or
-#'     centering within clusters. The regression coefficient of a de-meaned
-#'     variable represents the *within-subject-effect*.
-#'   }
+#' @section Why mixed models are preferred over fixed effects models:
 #'
-#'   \subsection{De-meaning with continuous predictors}{
-#'     For continuous time-varying predictors, the recommendation is to include
-#'     both their de-meaned and group-meaned versions as fixed effects, but not
-#'     the raw (untransformed) time-varying predictors themselves. The de-meaned
-#'     predictor should also be included as random effect (random slope). In
-#'     regression models, the coefficient of the de-meaned predictors indicates
-#'     the within-subject effect, while the coefficient of the group-meaned
-#'     predictor indicates the between-subject effect.
-#'   }
+#' A mixed models approach can model the causes of endogeneity explicitly
+#' by including the (separated) within- and between-effects of time-varying
+#' fixed effects and including time-constant fixed effects. Furthermore,
+#' mixed models also include random effects, thus a mixed models approach
+#' is superior to classic fixed-effects models, which lack information of
+#' variation in the group-effects or between-subject effects. Furthermore,
+#' fixed effects regression cannot include random slopes, which means that
+#' fixed effects regressions are neglecting "cross-cluster differences in the
+#' effects of lower-level controls (which) reduces the precision of estimated
+#' context effects, resulting in unnecessarily wide confidence intervals and
+#' low statistical power" (_Heisig et al. 2017_).
 #'
-#'   \subsection{De-meaning with binary predictors}{
-#'     For binary time-varying predictors, there are two recommendations. First
-#'     is to include the raw (untransformed) binary predictor as fixed effect
-#'     only and the *de-meaned* variable as random effect (random slope).
-#'     The alternative would be to add the de-meaned version(s) of binary
-#'     time-varying covariates as additional fixed effect as well (instead of
-#'     adding it as random slope). Centering time-varying binary variables to
-#'     obtain within-effects (level 1) isn't necessary. They have a sensible
-#'     interpretation when left in the typical 0/1 format (\cite{Hoffmann 2015,
-#'     chapter 8-2.I}). `demean()` will thus coerce categorical time-varying
-#'     predictors to numeric to compute the de- and group-meaned versions for
-#'     these variables, where the raw (untransformed) binary predictor and the
-#'     de-meaned version should be added to the model.
-#'   }
+#' @section Terminology:
 #'
-#'   \subsection{De-meaning of factors with more than 2 levels}{
-#'     Factors with more than two levels are demeaned in two ways: first, these
-#'     are also converted to numeric and de-meaned; second, dummy variables
-#'     are created (binary, with 0/1 coding for each level) and these binary
-#'     dummy-variables are de-meaned in the same way (as described above).
-#'     Packages like \pkg{panelr} internally convert factors to dummies before
-#'     demeaning, so this behaviour can be mimicked here.
-#'   }
+#' The group-meaned variable is simply the mean of an independent variable
+#' within each group (or id-level or cluster) represented by `by`. It represents
+#' the cluster-mean of an independent variable. The regression coefficient of a
+#' group-meaned variable is the *between-subject-effect*. The de-meaned variable
+#' is then the centered version of the group-meaned variable. De-meaning is
+#' sometimes also called person-mean centering or centering within clusters.
+#' The regression coefficient of a de-meaned variable represents the
+#' *within-subject-effect*.
 #'
-#'   \subsection{De-meaning interaction terms}{ There are multiple ways to deal
-#'   with interaction terms of within- and between-effects. A classical approach
-#'   is to simply use the product term of the de-meaned variables (i.e.
-#'   introducing the de-meaned variables as interaction term in the model
-#'   formula, e.g. `y ~ x_within * time_within`). This approach, however,
-#'   might be subject to bias (see \cite{Giesselmann & Schmidt-Catran 2020}).
-#'     \cr \cr
-#'     Another option is to first calculate the product term and then apply the
-#'     de-meaning to it. This approach produces an estimator \dQuote{that reflects
-#'     unit-level differences of interacted variables whose moderators vary
-#'     within units}, which is desirable if *no* within interaction of
-#'     two time-dependent variables is required. \cr \cr
-#'     A third option, when the interaction should result in a genuine within
-#'     estimator, is to "double de-mean" the interaction terms
-#'     (\cite{Giesselmann & Schmidt-Catran 2018}), however, this is currently
-#'     not supported by `demean()`. If this is required, the `wmb()`
-#'     function from the \pkg{panelr} package should be used. \cr \cr
-#'     To de-mean interaction terms for within-between models, simply specify
-#'     the term as interaction for the `select`-argument, e.g.
-#'     `select = "a*b"` (see 'Examples').
-#'   }
+#' @section De-meaning with continuous predictors:
 #'
-#'   \subsection{Analysing panel data with mixed models using lme4}{
-#'     A description of how to translate the
-#'     formulas described in *Bell et al. 2018* into R using `lmer()`
-#'     from \pkg{lme4} can be found in
-#'     [this vignette](https://easystats.github.io/parameters/articles/demean.html).
-#'   }
+#' For continuous time-varying predictors, the recommendation is to include
+#' both their de-meaned and group-meaned versions as fixed effects, but not
+#' the raw (untransformed) time-varying predictors themselves. The de-meaned
+#' predictor should also be included as random effect (random slope). In
+#' regression models, the coefficient of the de-meaned predictors indicates
+#' the within-subject effect, while the coefficient of the group-meaned
+#' predictor indicates the between-subject effect.
+#'
+#' @section De-meaning with binary predictors:
+#'
+#' For binary time-varying predictors, there are two recommendations. First
+#' is to include the raw (untransformed) binary predictor as fixed effect
+#' only and the *de-meaned* variable as random effect (random slope).
+#' The alternative would be to add the de-meaned version(s) of binary
+#' time-varying covariates as additional fixed effect as well (instead of
+#' adding it as random slope). Centering time-varying binary variables to
+#' obtain within-effects (level 1) isn't necessary. They have a sensible
+#' interpretation when left in the typical 0/1 format (_Hoffmann 2015,
+#' chapter 8-2.I_). `demean()` will thus coerce categorical time-varying
+#' predictors to numeric to compute the de- and group-meaned versions for
+#' these variables, where the raw (untransformed) binary predictor and the
+#' de-meaned version should be added to the model.
+#'
+#' @section De-meaning of factors with more than 2 levels:
+#'
+#' Factors with more than two levels are demeaned in two ways: first, these
+#' are also converted to numeric and de-meaned; second, dummy variables
+#' are created (binary, with 0/1 coding for each level) and these binary
+#' dummy-variables are de-meaned in the same way (as described above).
+#' Packages like **panelr** internally convert factors to dummies before
+#' demeaning, so this behaviour can be mimicked here.
+#'
+#' @section De-meaning interaction terms:
+#'
+#' There are multiple ways to deal with interaction terms of within- and
+#' between-effects.
+#'
+#' - A classical approach is to simply use the product term of the de-meaned
+#'   variables (i.e. introducing the de-meaned variables as interaction term
+#'   in the model formula, e.g. `y ~ x_within * time_within`). This approach,
+#'   however, might be subject to bias (see _Giesselmann & Schmidt-Catran 2020_).
+#'
+#' - Another option is to first calculate the product term and then apply the
+#'   de-meaning to it. This approach produces an estimator "that reflects
+#'   unit-level differences of interacted variables whose moderators vary
+#'   within units", which is desirable if *no* within interaction of
+#'   two time-dependent variables is required. This is what `demean()` does
+#'   internally when `select` contains interaction terms.
+#'
+#' - A third option, when the interaction should result in a genuine within
+#'   estimator, is to "double de-mean" the interaction terms
+#'   (_Giesselmann & Schmidt-Catran 2018_), however, this is currently
+#'   not supported by `demean()`. If this is required, the `wmb()`
+#'   function from the **panelr** package should be used.
+#'
+#' To de-mean interaction terms for within-between models, simply specify
+#' the term as interaction for the `select`-argument, e.g. `select = "a*b"`
+#' (see 'Examples').
+#'
+#' @section De-meaning for cross-classified designs:
+#'
+#' `demean()` can also handle cross-classified designs, where the data has two
+#' or more groups at the higher (i.e. second) level. In such cases, the
+#' `by`-argument can identify two or more variables that represent the
+#'  cross-classified group- or cluster-IDs. The de-meaned variables for
+#' cross-classified designs are simply subtracting all group means from each
+#' individual value, i.e. _fully cluster-mean-centering_ (see _Guo et al. 2024_
+#' for details). Note that de-meaning for cross-classified designs is *not*
+#' equivalent to de-meaning of nested data structures from models with three or
+#' more levels, i.e. de-meaning is supposed to work for models like
+#' `y ~ x + (1|group1) + (1|group2)`, but *not* for models like
+#' `y ~ x + (1|group1/group2)`.
+#'
+#' @section Analysing panel data with mixed models using lme4:
+#'
+#' A description of how to translate the formulas described in *Bell et al. 2018*
+#' into R using `lmer()` from **lme4** can be found in
+#' [this vignette](https://easystats.github.io/parameters/articles/demean.html).
 #'
 #' @references
 #'
@@ -185,6 +202,10 @@
 #'   - Giesselmann M, Schmidt-Catran, AW. 2020. Interactions in fixed
 #'   effects regression models. Sociological Methods & Research, 1–28.
 #'   https://doi.org/10.1177/0049124120914934
+#'
+#'   - Guo Y, Dhaliwal J, Rights JD. 2024. Disaggregating level-specific effects
+#'   in cross-classified multilevel models. Behavior Research Methods, 56(4),
+#'   3023–3057.
 #'
 #'   - Heisig JP, Schaeffer M, Giesecke J. 2017. The Costs of Simplicity:
 #'   Why Multilevel Models May Benefit from Accounting for Cross-Cluster
@@ -274,7 +295,8 @@ degroup <- function(x,
   center <- match.arg(tolower(center), choices = c("mean", "median", "mode", "min", "max"))
 
   if (inherits(select, "formula")) {
-    # formula to character, remove "~", split at "+"
+    # formula to character, remove "~", split at "+". We don't use `all.vars()`
+    # here because we want to keep the interaction terms as they are
     select <- trimws(unlist(
       strsplit(gsub("~", "", insight::safe_deparse(select), fixed = TRUE), "+", fixed = TRUE),
       use.names = FALSE
@@ -285,9 +307,12 @@ degroup <- function(x,
     by <- all.vars(by)
   }
 
+  # identify interaction terms
   interactions_no <- select[!grepl("(\\*|\\:)", select)]
   interactions_yes <- select[grepl("(\\*|\\:)", select)]
 
+  # if we have interaction terms that should be de-meaned, calculate the product
+  # of the terms first, then demean the product
   if (length(interactions_yes)) {
     interaction_terms <- lapply(strsplit(interactions_yes, "*", fixed = TRUE), trimws)
     product <- lapply(interaction_terms, function(i) do.call(`*`, x[, i]))
@@ -296,19 +321,21 @@ degroup <- function(x,
     select <- c(interactions_no, colnames(new_dat))
   }
 
-  not_found <- setdiff(select, colnames(x))
+  # check if all variables are present
+  not_found <- setdiff(c(select, by), colnames(x))
 
-  if (length(not_found) && isTRUE(verbose)) {
-    insight::format_alert(
-      sprintf(
-        "%i variables were not found in the dataset: %s\n",
-        length(not_found),
-        toString(not_found)
-      )
+  if (length(not_found)) {
+    insight::format_error(
+      paste0(
+        "Variable",
+        ifelse(length(not_found) > 1, "s ", " "),
+        text_concatenate(not_found, enclose = "\""),
+        ifelse(length(not_found) > 1, " were", " was"),
+        " not found in the dataset."
+      ),
+      .misspelled_string(colnames(x), not_found, "Possibly misspelled or not yet defined?")
     )
   }
-
-  select <- intersect(colnames(x), select)
 
   # get data to demean...
   dat <- x[, c(select, by)]
@@ -366,37 +393,60 @@ degroup <- function(x,
     max = function(.gm) max(.gm, na.rm = TRUE),
     function(.gm) mean(.gm, na.rm = TRUE)
   )
-  x_gm_list <- lapply(select, function(i) {
-    stats::ave(dat[[i]], dat[[by]], FUN = gm_fun)
-  })
-  names(x_gm_list) <- select
 
-  # create de-meaned variables by subtracting the group mean from each individual value
+  # we allow disaggregating level-specific effects for cross-classified multilevel
+  # models (see Guo et al. 2024). Two levels should work as proposed by the authors,
+  # more levels also already work, but need to check the formula from the paper
+  # and validate results
 
-  x_dm_list <- lapply(select, function(i) dat[[i]] - x_gm_list[[i]])
-  names(x_dm_list) <- select
+  if (length(by) == 1) {
+    # simple case: one level
+    group_means_list <- lapply(select, function(i) {
+      stats::ave(dat[[i]], dat[[by]], FUN = gm_fun)
+    })
+    names(group_means_list) <- select
+    # create de-meaned variables by subtracting the group mean from each individual value
+    person_means_list <- lapply(select, function(i) dat[[i]] - group_means_list[[i]])
+  } else {
+    # cross-classified design: by > 1
+    group_means_list <- lapply(by, function(j) {
+      out <- lapply(select, function(i) {
+        stats::ave(dat[[i]], dat[[j]], FUN = gm_fun)
+      })
+      names(out) <- paste0(select, "_", j)
+      out
+    })
+    # de-meaned variables for cross-classified design is simply subtracting
+    # all group means from each individual value
+    person_means_list <- lapply(seq_along(select), function(i) {
+      sum_group_means <- do.call(`+`, lapply(group_means_list, function(j) j[[i]]))
+      dat[[select[i]]] - sum_group_means
+    })
+  }
 
+  # preserve names
+  names(person_means_list) <- select
 
   # convert to data frame and add suffix to column names
 
-  x_gm <- as.data.frame(x_gm_list)
-  x_dm <- as.data.frame(x_dm_list)
+  group_means <- as.data.frame(group_means_list)
+  person_means <- as.data.frame(person_means_list)
 
-  colnames(x_dm) <- sprintf("%s%s", colnames(x_dm), suffix_demean)
-  colnames(x_gm) <- sprintf("%s%s", colnames(x_gm), suffix_groupmean)
+  colnames(person_means) <- sprintf("%s%s", colnames(person_means), suffix_demean)
+  colnames(group_means) <- sprintf("%s%s", colnames(group_means), suffix_groupmean)
 
   if (isTRUE(add_attributes)) {
-    x_dm[] <- lapply(x_dm, function(i) {
+    person_means[] <- lapply(person_means, function(i) {
       attr(i, "within-effect") <- TRUE
       i
     })
-    x_gm[] <- lapply(x_gm, function(i) {
+    group_means[] <- lapply(group_means, function(i) {
       attr(i, "between-effect") <- TRUE
       i
     })
   }
 
-  cbind(x_gm, x_dm)
+  cbind(group_means, person_means)
 }