+++ NEWS.md
@@ -1,4 +1,4 @@
-# datawizard 0.11.0.1
+# datawizard 0.12.0
 
 BREAKING CHANGES